+++ dev/PyRFM/Utilities/Permutation.py
@@ -1,0 +1,9 @@
+import numpy as np
+
+
+def new_permutation(length, rng):
+    perm = np.arange(length)
+    rng.shuffle(perm)
+    iperm = np.argsort(perm)
+
+    return perm, iperm
